requirements.txt
torch==1.12.0
PyYAML==5.4.1
transformers==4.22.1
tokenizers==0.12.0
scikit-learn==1.1.2
pandas==1.4.2
datasets==2.4.0
matplotlib==3.5.2
protobuf==3.20.1
